gateway-service/app/requirements.txt
requests
pyjwt[crypto]==2.6.0
